trac:3995 - PA Wifi script pa_wifi.sh is not runnable
Patch3:
- fixes on path for Shell script used in WifiService. Warning: Wifi Client & AP sample applications do not start correctly on Legato 16.01.

Patch2:
- Commit incomplete for script path fix

Patch1:
- Adaptations for Legato 16.01 release compatibility

+++ apps/sample/wifiClientTest/wifiClientTestComponent/wifiClientTest.c
@@ -70,8 +70,8 @@ static void TestReadScanResults
             {
                 LE_DEBUG( "TestReadScanResults:le_wifiClient_GetSsid OK, ssidLength %d;"
                                     "SSID: \"%.*s\" ",
-                                    ssidNumElements,
-                                    ssidNumElements,
+                                    (int) ssidNumElements,
+                                    (int) ssidNumElements,
                                     (char*) &ssidBytes[0]);
         }
         else
